Convert QuadConverter to use ES6 modules
