+++ app/assets/javascripts/student_profile/profile_chart.jsx
@@ -1,11 +1,10 @@
 import _ from 'lodash';
 import {merge} from '../helpers/react_helpers.jsx';
 import HighchartsWrapper from '../student_profile/HighchartsWrapper.js';
+import QuadConverter from '../student_profile/QuadConverter.js';
 
 (function() {
   window.shared || (window.shared = {});
-  const QuadConverter = window.shared.QuadConverter;
-
   const ProfileChartSettings = window.ProfileChartSettings;
 
   // Component for all charts in the profile page.
