+++ ui/legacy.js
@@ -8,7 +8,6 @@ import './legacyHelpers';
 
 // student profile page:
   // pure ui components:
-import '../app/assets/javascripts/student_profile/quad_converter.jsx';
 import '../app/assets/javascripts/student_profile/provided_by_educator_dropdown.jsx';
 import '../app/assets/javascripts/student_profile/profile_chart_settings.jsx';
 import '../app/assets/javascripts/student_profile/bar_chart_sparkline.jsx';
